Use IServiceProvider directly in middleware constructor
The middleware class had previously taken an collection of health check runners in its constructor, and if the collection was empty, the static HealthCheck class would be used to obtain the middleware's health check runner. By using the IServiceProvider directly, we first attempt to resolve a collection of health check runners directly. If none can be resolved, we then fall back to configuration, similar to before. However, a "better" configuration is used - one resolved from the service provider (which ASP.NET Core registers for the app). Another benefit is that ConfigurationObjectFactory is able to use the service provider to resolve any dependencies not defined in configuration.